# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: E-Logbook Enterprise API
  description: |
    API for the E-Logbook Enterprise platform — a medical resident case-logging
    and faculty evaluation system with multi-tenant support, SSO, AI-powered
    educational insights, payment integration, and SCIM user provisioning.

    ## Authentication
    Most endpoints require JWT authentication via Supabase Auth. Pass the access
    token as a Bearer token in the `Authorization` header. Public endpoints
    (health, SSO check) do not require authentication.

    Edge Functions use Supabase's built-in JWT verification through the
    shared `authenticate()` helper.
  version: "1.0.0"
  contact:
    name: E-Logbook Support
    url: https://elogbook.dev
servers:
  - url: https://app.elogbook.dev
    description: Production server
  - url: http://localhost:3000
    description: Local development
  - url: https://{tenant}.elogbook.dev
    description: Tenant-specific subdomain
    variables:
      tenant:
        default: demo
        description: Tenant slug

# ============================================================================
# Components
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token
    ServiceRole:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase service_role key (admin operations)
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true

    # --------------- Health ---------------
...
    AiConfigRequest:
      type: object
      required: [model]
      properties:
        provider:
          type: string
          enum: [openai, openrouter, anthropic, azure, aihubmix, custom]
          description: AI provider name
        model:
          type: string
          description: Model identifier (e.g., gpt-4, claude-3-opus)
        is_active:
          type: boolean
          default: false
        endpoint_url:
          type: string
          format: uri
          description: Custom endpoint URL (for Azure/custom providers)
        api_key:
          type: string
          description: API key for the provider (required for new configs)
    AiConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        has_key:
          type: boolean

    # --------------- Assign Role ---------------
...
    PaymentGatewayRequest:
      type: object
      required: [publishable_key]
      properties:
        provider:
          type: string
          enum: [stripe]
          default: stripe
        publishable_key:
          type: string
        is_active:
          type: boolean
          default: false
        endpoint_url:
          type: string
          format: uri
        secret_key:
          type: string
        webhook_secret:
          type: string
    PaymentGatewayResponse:
      $ref: '#/components/schemas/Success'

    # --------------- SSO Config ---------------
...
paths:
...
  /api/{tenant}/admin/ai-config:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
        description: Tenant slug
    post:
      operationId: upsertAiConfig
      summary: Create or update AI provider configuration
      description: |
        Configure (create or update) the AI provider settings for a tenant.
        Requires institution_admin or admin role. CSRF protected.
      tags: [Admin]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiConfigRequest'
      responses:
        "200":
          description: Configuration saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiConfigResponse'
        "400":
          description: Validation error (missing model, etc.)
        "401":
          description: Unauthorized
        "403":
          description: Insufficient permissions or tenant mismatch
        "429":
          description: Rate limited
        "500":
          description: Internal server error
    put:
      operationId: updateAiConfig
      summary: Update AI provider configuration (alias for POST)
      tags: [Admin]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiConfigRequest'
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiConfigResponse'

  # ============= Assign Role =============
...
  /api/{tenant}/admin/payment-gateway:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: upsertPaymentGateway
      summary: Create or update payment gateway configuration
      description: |
        Configure Stripe payment gateway for a tenant. Keys are encrypted
        at rest. Requires institution_admin or admin role.
      tags: [Admin, Billing]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentGatewayRequest'
      responses:
        "200":
          description: Gateway configuration saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentGatewayResponse'
        "400":
          description: Missing publishable key
        "401":
          description: Unauthorized
        "403":
          description: Insufficient permissions
        "429":
          description: Rate limited
    put:
      operationId: updatePaymentGateway
      summary: Update payment gateway configuration (alias for POST)
      tags: [Admin, Billing]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentGatewayRequest'
      responses:
        "200":
          description: Gateway configuration updated

  # ============= SSO Config =============
...
  /api/{tenant}/compliance/export:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: exportComplianceReport
      summary: Export compliance report
      description: |
        Exports compliance data (data-access, PHI inventory, consent
        records, or retention status) as CSV or HTML. Requires director,
        institution_admin, or admin role. Rate limited to 10 requests
        per minute.
      tags: [Compliance]
      security:
        - BearerAuth: []
      parameters:
        - name: section
          in: query
          required: true
          schema:
            type: string
            enum: [data-access, phi-inventory, consent, retention]
          description: Compliance section to export
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, pdf]
            default: csv
      responses:
        "200":
          description: Compliance export file
          content:
            text/csv:
              schema:
                type: string
            text/html:
              schema:
                type: string
        "400":
          description: Invalid section or format
        "401":
          description: Unauthorized
        "403":
          description: Insufficient permissions
        "429":
          description: Rate limited

  # ============= Export PDF =============
...